+++ examples/shared/xy-candlestick/index.ts
@@ -56,15 +56,21 @@ var chart = root.container.children.push(
 var xAxis = chart.xAxes.push(
   am5xy.DateAxis.new(root, {
     groupData: true,
+    maxDeviation:1,
     baseInterval: { timeUnit: "day", count: 1 },
-    renderer: am5xy.AxisRendererX.new(root, {}),
+    renderer: am5xy.AxisRendererX.new(root, {
+      pan:"zoom"
+    }),
     tooltip: am5.Tooltip.new(root, {})
   })
 );
 
 var yAxis = chart.yAxes.push(
   am5xy.ValueAxis.new(root, {
-    renderer: am5xy.AxisRendererY.new(root, {})
+    maxDeviation:1,
+    renderer: am5xy.AxisRendererY.new(root, {
+      pan:"zoom"
+    })
   })
 );